+++ backend/routes/api/imagens.js
@@ -1,10 +1,11 @@
 const express = require('express');
 const router = express.Router();
 const db = require('../../models');
-const upImag = require('../../upImag'); // importa tu lógica de upload
+const multer = require('multer');
+const upload = multer({ dest: 'uploads/' });
 
-// Rota para upload de imagem (ajusta según tu lógica en upImag.js)
-router.post('/upload', upImag.single('imagem'), async (req, res) => {
+// Rota para upload de imagem
+router.post('/upload', upload.single('imagem'), async (req, res) => {
   try {
     const { nome, valor, categoria, isOferta } = req.body;
     const imagemUrl = req.file ? req.file.path : null;
