+++ frontend/src/components/carousel.js
