+++ frontend/src/components/footer.js
