+++ frontend/src/pages/Registro.js
