+++ frontend/src/pages/subpages/Lojas.js
