--- projeto-final-fullstack-main/backend/models/Usuario.js
@@ -1,24 +1,0 @@
-// Modelo Sequelize para a tabela 'usuarios'
-
-const { DataTypes } = require('sequelize');
-const { sequelize } = require('./index'); // importa a conexão
-
-// Modelo Sequelize para a tabela 'usuarios'
-const Usuario = sequelize.define('Usuario', {
-  nome: {
-    type: DataTypes.STRING,
-    allowNull: false,
-  },
-  email: {
-    type: DataTypes.STRING,
-    allowNull: false,
-    unique: true,
-  },
-  comportamento: {
-    type: DataTypes.STRING,
-  }
-}, {
-  tableName: 'usuarios'
-});
-
-module.exports = Usuario;
